Move MolecularOrbitals to its own module
